==> 1. Amenities_cycleways_extraction.ipynb <==
import osmnx as ox
import requests
import pandas as pd
import geopandas as gpd
import os
import numpy as np
import warnings
import time
import math

%%time
city = 'Paris'
warnings.filterwarnings('ignore')
# Extract bike networks according to this list
Bike_ParisC = ox.graph_from_place(['Paris','Saint-Mandé, Ile-de-France'], network_type = "bike", 
                                            retain_all = True, clean_periphery = True)

%%time
# Get the nodes and edges from the graph
nodes, edges = ox.graph_to_gdfs(Bike_ParisC)

%%time
# Find cycleway if present
edges['highway'] = np.where(edges['highway'].isin(['cycleway']),'cycleway',edges['highway'])

# Unlist items if needed. First item (as most important) counts
for j in range(len(edges.columns)):
    for i in range(len(edges)):
        if isinstance(edges.iloc[i,j], list):
            edges.iloc[i,j] = edges.iloc[i,j][0]
edges

# Removing roads equal to or faster than 90 km/h
ed2 = edges[~edges['maxspeed'].isin(['walk','FR:walk','signals'])]
ed2 = pd.concat([ed2[(ed2['maxspeed'].astype(float) >= 90) == False],
                 edges[edges['maxspeed'].isin(['walk','FR:walk','signals'])]])

# Make a dedicated "number-based" speed column
speed = ed2['maxspeed']
speed = pd.Series(np.where(speed.isin(['walk','FR:walk']),10,np.where(speed.isin(['signals']),50,speed)), dtype = float)
speed.index = ed2.index

# Transform oneway and lanes to numberwise columns
ed2['cleanspeed'] = speed
ed2['oneway'] = ed2['oneway'].astype(int)
ed2['lanes'] = pd.Series(np.where(ed2['lanes'] =='Allée des petits clos',1,ed2['lanes']),index = ed2.index, dtype =float)
ed2

%%time
# Set nodes location to meters (for later filter in R)
pd.set_option('display.float_format', lambda x: '%.1f' % x)
nodes['geom_x_m'] = nodes['geometry'].to_crs(3043).x
nodes['geom_y_m'] = nodes['geometry'].to_crs(3043).y

nodes['osmid_var'] = nodes.index
nodes['osmid_var'] = nodes['osmid_var'].astype('int64')

# Get edge coordinates
c1 = list()
c2 = list()
c3 = list()
c4 = list()
geom = ed2.to_crs(3043).geometry
for i in range(len(ed2)):
    l1 = list(geom.iloc[i].coords)
    c1.append(l1[0][0])
    c2.append(l1[0][1])
    c3.append(l1[-1][0])
    c4.append(l1[-1][1])

# Add edge coordinates to df
ed2['x1_m'] = c1
ed2['y1_m'] = c2
ed2['x2_m'] = c3
ed2['y2_m'] = c4

# Make if path not exists and save the files there.
warnings.filterwarnings('ignore')
if not os.path.exists('D:/EconNet/'+str(city)+'/Cycling_City/'):
    os.makedirs('D:/EconNet/'+str(city)+'/Cycling_City/')
nodes.to_file('D:/EconNet/'+str(city)+'/Cycling_City/BikePoints.shp')
ed2.to_file('D:/EconNet/'+str(city)+'/Cycling_City/BikeRoutes.shp')
nodes.sort_values('osmid')

# Extract transit access points and bike stations from Apur (open source data Ile-de-France)
warnings.filterwarnings('ignore')
transit_access = gpd.read_file(r"D:\EconNet\acces.geojson")
bike_stations = gpd.read_file(r"D:\EconNet\velib-emplacement-des-stations.geojson")

# Take the Paris boundaries
Paris_bounds = ox.geocode_to_gdf(['Paris','Saint-Mandé, Ile-de-France'])
Paris_bounds

warnings.filterwarnings('ignore')

# Overlay city boundaries with the bike stations, only include ones inside the selected area
Paris_transitaccess = transit_access.overlay(Paris_bounds)

# Format to fit in df of amenities, take geoms x-y coordinates.
Paris_transitaccess['name'] = Paris_transitaccess['fournisseurname']+'-'+Paris_transitaccess['accname']
Paris_transitaccess['amenity'] = 'transit_entry'
Paris_transitaccess['osmid'] = Paris_transitaccess['accid']
Paris_transitaccess = Paris_transitaccess[['osmid','amenity','name','geometry']]
Paris_transitaccess['geometry'] = Paris_transitaccess.centroid
Paris_transitaccess['geom_x_m'] = Paris_transitaccess['geometry'].to_crs(3043).x
Paris_transitaccess['geom_y_m'] = Paris_transitaccess['geometry'].to_crs(3043).y
Paris_transitaccess.to_file('D:/EconNet/Paris/POI_City/OSM_transit_entry.shp')
Paris_transitaccess

warnings.filterwarnings('ignore')

# Overlay city boundaries with the bike stations, only include ones inside the selected area
Paris_bikestations = bike_stations.overlay(Paris_bounds)

# Format to fit in df of amenities, take geoms x-y coordinates.
Paris_bikestations['amenity'] = 'bikestation'
Paris_bikestations['osmid'] = Paris_bikestations['stationcode']
Paris_bikestations = Paris_bikestations[['osmid','amenity','name','geometry']]
Paris_bikestations['geometry'] = Paris_bikestations.centroid
Paris_bikestations['geom_x_m'] = Paris_bikestations['geometry'].to_crs(3043).x
Paris_bikestations['geom_y_m'] = Paris_bikestations['geometry'].to_crs(3043).y
Paris_bikestations.to_file('D:/EconNet/Paris/POI_City/OSM_bikestations.shp')
Paris_bikestations

%%time
# suppress scientific notation of geometries
pd.set_option('display.float_format', lambda x: '%.1f' % x)
start_time = time.time()

# Amenities and shops that need to be extracted
s = ['restaurant','cafe','college','university','pharmacy','dentist','doctors','cinema','theatre','supermarket',
            'bakery','butcher','greengrocer']
m = ['amenity','amenity','amenity','amenity','amenity','amenity','amenity','amenity','amenity','shop','shop','shop','shop']
Paris_amenities = pd.DataFrame()

if not os.path.exists('D:/EconNet/'+str(city)+'/POI_City/'):
    os.makedirs('D:/EconNet/'+str(city)+'/POI_City/')
    
# For every amenity/shop, extract from OSM
for i in range(len(s)):
    warnings.filterwarnings('ignore')
    locals()[s[i]] = ox.geometries_from_place(['Paris','Saint-Mandé, Ile-de-France'], tags={m[i]:s[i]})
    
    # Subset and rename columns
    locals()[s[i]] = locals()[s[i]][[m[i],'name','geometry']]
    locals()[s[i]].columns = ['amenity','name','geometry']
    print('path: D:/EconNet/'+str(city)+'/POI_City/OSM_'+m[i]+'_'+s[i]+'.shp')
    
    # Take the centroid of the location for simpler computation in R
    warnings.filterwarnings('ignore')
    locals()[s[i]]['geometry'] = locals()[s[i]]['geometry'].centroid
    
    # Get X and Y coordinates for computation in R
    locals()[s[i]]['geom_x_m'] = locals()[s[i]]['geometry'].to_crs(3043).x
    locals()[s[i]]['geom_y_m'] = locals()[s[i]]['geometry'].to_crs(3043).y
    
    # Save the specific amenity as shapefile
    locals()[s[i]].to_file('D:/EconNet/'+str(city)+'/POI_City/OSM_'+m[i]+'_'+s[i]+'.shp')
    
    # Add the amenity to the mastertable of amenities/shops
    Paris_amenities = pd.concat([Paris_amenities,locals()[s[i]]])
    print(s[i],'done',round((time.time() - start_time) / 60,2),' mns')

# Format and export the mastertable
Paris_amenities = pd.concat([Paris_amenities,Paris_bikestations,Paris_transitaccess])
Paris_amenities['element_ty'] = Paris_amenities['element_ty'].fillna('node')
Paris_amenities = Paris_amenities.reset_index()
Paris_amenities = Paris_amenities.iloc[:,2:]
Paris_amenities = Paris_amenities.drop_duplicates()
Paris_amenities.to_file('D:/EconNet/Paris/POI_City/amenities_shops.shp')
Paris_amenities
